pyproject.toml
[project]
name = "agent-command-center"
version = "0.1.2"
description = "A TUI to monitor and manage coding agent sessions (Claude, OpenCode, Codex, Aider, Gemini) running in tmux"
license = { text = "GNU AGPLv3" }
requires-python = ">=3.11"
dependencies = [
    "textual>=0.89",
    "anthropic>=0.40",
    "pyyaml>=6.0",
    "psutil>=5.9",
]

[project.scripts]
acc = "acc.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/acc"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
